Move Subsector to its own class
