+++ PyRoute/DeltaDebug/DeltaGalaxy.py
@@ -5,8 +5,9 @@ Created on May 21, 2023
 """
 
 from PyRoute.AreaItems.Allegiance import Allegiance
+from PyRoute.AreaItems.Subsector import Subsector
 from PyRoute.DeltaDebug.DeltaDictionary import SectorDictionary, SubsectorDictionary
-from PyRoute.Galaxy import Galaxy, Sector, Subsector
+from PyRoute.Galaxy import Galaxy, Sector
 from PyRoute.Star import Star
 from PyRoute.AllyGen import AllyGen
